Suppress overflow warning in _LoadData.py
While working with hourly building load data, I repeatedly encountered a `RuntimeWarning` about an overflow in `float16` which worried me. After looking into the code, I saw that the warning is harmless/not necessary: if an overflow occurs, the array is automatically recreated using `float32` in the next lines. To avoid confusion, I suggest to suppress the warning.